requirements.txt
kafka-python
sseclient-py
requests
redis
fastapi
uvicorn[standard]
rich
